Guidelines for efficient DApps
This is a bulletpoints draft with the main topics to be covered. PLEASE squash or remove this commit once the final content is added.

Topics include

Recommendations for an efficient way of using Inspect+GraphQL APIs, so that DApps can support a large number of clients.
General recommendation of preferring compiled languages (C++ and Rust) instead of interpreted languages (Javascript/Typescript, Python, Lua) when developing "serious" DApps, especially if performance is a concern
Closes #108